ai_modules/backend/requirements.txt
# FastAPI 관련
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
pydantic==1.8.2

# 웹소켓
websockets==10.0

# AI/ML 관련
google-cloud-aiplatform==1.16.0
google-cloud-storage==1.42.0
numpy==1.21.2
pandas==1.3.3

# 유틸리티
python-multipart==0.0.5
aiofiles==0.7.0
python-jose==3.3.0
passlib==1.7.4

# 데이터베이스
sqlalchemy==1.4.23
alembic==1.7.1

# 테스트
pytest==6.2.5
pytest-asyncio==0.15.1